requirements.txt
torch==2.10.0 
torchvision==0.25.0 
transformers==5.2.0 
pillow==12.0.0 
opencv-python==4.13.0.92 
accelerate==1.12.0 
pynput==1.8.1 
piper-tts==1.4.1 
sounddevice==0.5.5 
numpy==2.3.5 
onnxruntime==1.24.2 